sensehat_aca/requirements.txt
flask==3.1.0
azure-cosmos==4.7.0
gunicorn==23.0.0
